move core xarray support to _core
it's being used in many places in _core already
